Medical_IoT_System/analytics-python/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
requests
pycryptodome
